+++ src/components/Download.tsx
@@ -1,21 +1,8 @@
 import { Box, Container, Heading, Text, VStack, Stack, useColorModeValue, Flex, Icon } from '@chakra-ui/react'
 import { FaApple } from 'react-icons/fa'
 import { IoLogoGooglePlaystore } from 'react-icons/io5'
-import { useThemedToast } from '../lib/useThemedToast'
 
 const Download = () => {
-  const toast = useThemedToast()
-
-  const handleAndroidClick = (e: React.MouseEvent) => {
-    e.preventDefault()
-    toast({
-      title: "Coming Soon!",
-      description: "SpySocial will be available on Android soon. Stay tuned!",
-      status: "info",
-      duration: 3000,
-    })
-  }
-
   return (
     <Box 
       id="download" 
@@ -69,8 +56,9 @@ const Download = () => {
             
             <Box
               as="a"
-              href="#"
-              onClick={handleAndroidClick}
+              href="https://play.google.com/store/apps/details?id=com.gerz.spysocial"
+              target="_blank"
+              rel="noopener noreferrer"
               display="inline-flex"
               alignItems="center"
               justifyContent="center"